# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:3220
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]
    OrderItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
          maxItems: 50
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
          maxItems: 50
    OrderItemResponse:
      type: object
      required: [id, product_id, quantity, price_at_order]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: double
    OrderResponse:
      type: object
      required: [id, user_id, status, total_amount, discount_amount, items, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code_id:
          type: string
          format: uuid
          nullable: true
        total_amount:
          type: number
          format: double
        discount_amount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          enum:
            - PRODUCT_NOT_FOUND
            - PRODUCT_INACTIVE
            - ORDER_NOT_FOUND
            - ORDER_LIMIT_EXCEEDED
            - ORDER_HAS_ACTIVE
            - INVALID_STATE_TRANSITION
            - INSUFFICIENT_STOCK
            - PROMO_CODE_INVALID
            - PROMO_CODE_MIN_AMOUNT
            - ORDER_OWNERSHIP_VIOLATION
            - VALIDATION_ERROR
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - REFRESH_TOKEN_INVALID
            - ACCESS_DENIED
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
paths:
...
  /orders:
    post:
      tags: [orders]
      summary: Create order
      description: 'Roles: USER, ADMIN. SELLER - forbidden (ACCESS_DENIED). Rate-limited: ORDER_LIMIT_EXCEEDED if last CREATE_ORDER < N minutes ago. Forbidden if active order exists: ORDER_HAS_ACTIVE.'
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Promo code error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [orders]
      summary: Get order by ID
      description: 'Roles: USER (own orders only), ADMIN (any). SELLER - forbidden (ACCESS_DENIED).'
      operationId: getOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [orders]
      summary: Update order items
      description: 'Roles: USER (own orders only), ADMIN (any). SELLER - forbidden (ACCESS_DENIED). Only allowed in CREATED state, otherwise INVALID_STATE_TRANSITION. Rate-limited: ORDER_LIMIT_EXCEEDED if last UPDATE_ORDER < N minutes ago.'
      operationId: updateOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [orders]
      summary: Cancel order
      description: 'Roles: USER (own orders only), ADMIN (any). SELLER - forbidden (ACCESS_DENIED). Allowed only from CREATED or PAYMENT_PENDING states, otherwise INVALID_STATE_TRANSITION. Returns stock and decrements promo code uses.'
      operationId: cancelOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...